requirements.txt
discordwebhook
pandas
pytesseract
python-telegram-bot==13.7
PyYAML
requests
schedule
Pillow